Localiztion reworked. Relocated and renamed screen classes
